delete confirm dialog and user delete paths

+++ libs/wivipro/feat-manage/src/lib/users/offcanvas/user.offcanvas.ts
@@ -1,20 +1,24 @@
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, Validators } from '@angular/forms';
+import { Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import {
   AppUser,
   AuthApi,
   AuthUser,
   UserApi,
   NotificationService,
+  ModalService,
 } from '@geerts/shared';
 import { NgbActiveOffcanvas } from '@ng-bootstrap/ng-bootstrap';
-import { concatMap, Observable, of } from 'rxjs';
+import { concatMap, switchMap, Observable, of } from 'rxjs';
+import { NEVER } from 'rxjs';
 
 @Component({
   selector: 'geerts-user-offcanvas',
   templateUrl: './user.offcanvas.html',
 })
 export class UserOffCanvasComponent implements OnInit {
+  @ViewChild('confirmDeleteUser') confirmDeleteUser!: TemplateRef<Element>;
+
   usr!: AppUser;
   title!: string;
   action: 'create' | 'update' = 'create';
@@ -22,6 +26,7 @@ export class UserOffCanvasComponent implements OnInit {
   resultMessage: 'Gebruiker aangemaakt' | 'Gebruiker aangepast' =
     'Gebruiker aangemaakt';
   loading = false;
+  initialFormValues!: string[];
 
   form = this.fb.group({
     email: ['', Validators.email],
@@ -39,7 +44,8 @@ export class UserOffCanvasComponent implements OnInit {
     private fb: FormBuilder,
     private authApi: AuthApi,
     private userApi: UserApi,
-    private notificationService: NotificationService
+    private notificationService: NotificationService,
+    private modalService: ModalService
   ) {}
 
   ngOnInit(): void {
@@ -63,45 +69,50 @@ export class UserOffCanvasComponent implements OnInit {
       phoneNumber: this.usr.phoneNumber || '',
       password: '',
     });
+    // used to check changes later on
+    this.initialFormValues = this.toArray(this.form);
   }
 
   // TODO validate form, cause email validator is wrong
   ok(): void {
-    this.loading = this.notificationService.showLoading();
-    let $uid: Observable<string> = new Observable<string>();
-    const userFromForm: AppUser = this.formToUser();
-    if (this.action === 'create') {
-      // if the action is 'create a user'
-      // first we need to make an actual authentication record for the user
-      const newAuth: AuthUser = {
-        email: this.form.get('email')?.value,
-        password: this.form.get('password')?.value,
-        displayName: this.form.get('displayName')?.value,
-      };
-      $uid = this.authApi.create(newAuth);
+    if (this.formHasChanged(this.initialFormValues, this.toArray(this.form))) {
+      this.loading = this.notificationService.showLoading();
+      let $uid: Observable<string> = new Observable<string>();
+      const userFromForm: AppUser = this.formToUser();
+      if (this.action === 'create') {
+        // if the action is 'create a user'
+        // first we need to make an actual authentication record for the user
+        const newAuth: AuthUser = {
+          email: this.form.get('email')?.value,
+          password: this.form.get('password')?.value,
+          displayName: this.form.get('displayName')?.value,
+        };
+        $uid = this.authApi.create(newAuth);
+      } else {
+        // update existing user
+        $uid = of(userFromForm.uid);
+      }
+      $uid
+        .pipe(
+          concatMap((uid) => {
+            userFromForm.uid = uid;
+            return this.userApi.create(userFromForm);
+          })
+        )
+        .subscribe({
+          next: (r) => {
+            this.activeOffCanvas.close(r);
+            this.loading = this.notificationService.hideLoading();
+            this.notificationService.success(this.resultMessage);
+          },
+          error: (e) => {
+            this.loading = this.notificationService.hideLoading();
+            this.notificationService.handleApiError(e);
+          },
+        });
     } else {
-      // update existing user
-      $uid = of(userFromForm.uid);
+      this.cancel();
     }
-
-    $uid
-      .pipe(
-        concatMap((uid) => {
-          userFromForm.uid = uid;
-          return this.userApi.create(userFromForm);
-        })
-      )
-      .subscribe({
-        next: (r) => {
-          this.activeOffCanvas.close(r);
-          this.loading = this.notificationService.hideLoading();
-          this.notificationService.success(this.resultMessage);
-        },
-        error: (e) => {
-          this.loading = this.notificationService.hideLoading();
-          this.notificationService.handleApiError(e);
-        },
-      });
   }
 
   cancel(): void {
@@ -110,13 +121,20 @@ export class UserOffCanvasComponent implements OnInit {
 
   // TODO confirm delete
   delete(): void {
-    this.loading = this.notificationService.showLoading();
-    // first delete the user record, if that succeeds, delete the auth record as well
-    this.userApi
-      .delete(this.usr.uid)
+    this.modalService
+      .confirmDeleteWithTemplate(this.confirmDeleteUser)
       .pipe(
-        concatMap((res) => {
-          return this.authApi.delete(this.usr.uid);
+        switchMap((r) => {
+          if (r.Success) {
+            this.loading = this.notificationService.showLoading();
+            return this.userApi.delete(this.usr.uid).pipe(
+              concatMap((_) => {
+                return this.authApi.delete(this.usr.uid);
+              })
+            );
+          } else {
+            return NEVER;
+          }
         })
       )
       .subscribe({
@@ -144,5 +162,20 @@ export class UserOffCanvasComponent implements OnInit {
     };
     const ret: AppUser = { ...this.usr, ...newUsr };
     return ret;
+  }
+
+  private formHasChanged(
+    initialValues: string[],
+    newValues: string[]
+  ): boolean {
+    return initialValues.some((val, idx) => val !== newValues[idx]);
+  }
+
+  private toArray(form: FormGroup): string[] {
+    const arr = new Array<string>();
+    Object.keys(form.controls).forEach((key) => {
+      arr.push(form.get(key)?.value);
+    });
+    return arr;
   }
 }
